Fix plt.text() numpy scalar type errors
np.mean() returns a numpy.floating scalar, not a plain float. mypy
locally didn't flag this against the installed matplotlib stub
version, but CI's fresh install pulled a stricter one that rejects
numpy.floating where plt.text() expects float. Wrap both call sites
in float() to satisfy the stub regardless of stub version.

+++ eztfem/addons/plotlib/plotting_functions.py
@@ -426,8 +426,9 @@ def plot_mesh_plt(
     # Plot element numbers
     if elementnumbers:
         for elem in range(mesh.nelem):
-            plt.text(np.mean(x[:, elem]), np.mean(y[:, elem]), str(elem),
-                     fontsize=8, color='r', ha='center', va='center')
+            plt.text(float(np.mean(x[:, elem])), float(np.mean(y[:, elem])),
+                     str(elem), fontsize=8, color='r', ha='center',
+                     va='center')
 
     plt.show()
 
@@ -520,7 +521,8 @@ def plot_curves(
         for curve in curves:
             for elem in range(mesh.curves[curve].nelem):
                 el += 1
-                plt.text(np.mean(x[:, el-1]), np.mean(y[:, el-1]),
+                plt.text(float(np.mean(x[:, el-1])),
+                         float(np.mean(y[:, el-1])),
                          str(elem + 1), fontsize=8, color='r', ha='center',
                          va='center')
